requirements.txt
selenium==4.9.0
sendgrid
dummylog
python-dotenv